Removed lambda-attributes
