--- lambda-attributes/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "lambda-attributes"
-version = "0.1.0"
-authors = ["David Barsky <dbarsky@amazon.com>"]
-edition = "2018"
-license = "Apache-2.0"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = "1.0"
-syn = { version = "1.0.5", features = ["full"] }
-quote = "1"
