+++ lambda/examples/hello.rs
@@ -1,10 +1,15 @@
-use lambda::{lambda, Context};
+use lambda::{handler_fn, Context};
 use serde_json::Value;
 
 type Error = Box<dyn std::error::Error + Send + Sync + 'static>;
 
-#[lambda]
 #[tokio::main]
-async fn main(event: Value, _: Context) -> Result<Value, Error> {
+async fn main() -> Result<(), Error> {
+    let func = handler_fn(func);
+    lambda::run(func).await?;
+    Ok(())
+}
+
+async fn func(event: Value, _: Context) -> Result<Value, Error> {
     Ok(event)
 }
